IRIS-3963 | merge FacebookTests and FacebookButtonTests
